requirements.txt
Flask>=2.2
Flask-WTF>=1.0
Flask-Bcrypt>=1.0
Flask-SQLAlchemy>=3.0
gunicorn>=20.1
numpy>=1.24
redis>=4.5
pyodbc>=4.0
fasttext-wheel>=0.9.2
python-dotenv>=1.0
pytest>=7.0
requests>=2.30Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1
Flask-Limiter==3.7.0
pyodbc==5.1.0
fasttext-wheel==0.9.2
numpy==1.26.4
redis==5.0.4
python-dotenv==1.0.1
gunicorn==22.0.0
